Updates from Lesson 3

+++ conFusion/www/js/controllers.js
@@ -228,20 +228,35 @@ angular.module('conFusion.controllers', [])
 
                     }])
 
-    .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+        //This code got changed in Lesson3
+        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
             $scope.baseURL = baseURL;
             $scope.shouldShowDelete = false;
 
+            //New spinner from Lesson3
+            $ionicLoading.show({
+                    template: '<ion-spinner></ion-spinner> Loading...'
+            });
+
             $scope.favorites = favoriteFactory.getFavorites();
 
+            //Code update in Lesson3 to simulate spinner
             $scope.dishes = menuFactory.getDishes().query(
                 function (response) {
                     $scope.dishes = response;
+                    $timeout(function () {
+                        $ionicLoading.hide();
+                    }, 1000);
                 },
                 function (response) {
                     $scope.message = "Error: " + response.status + " " + response.statusText;
+                    $timeout(function () {
+                        $ionicLoading.hide();
+                    }, 1000);
                 });
+
+
             console.log($scope.dishes, $scope.favorites);
 
             $scope.toggleDelete = function () {
@@ -249,11 +264,27 @@ angular.module('conFusion.controllers', [])
                 console.log($scope.shouldShowDelete);
             }
 
+            //This code got changed in Lesson3
             $scope.deleteFavorite = function (index) {
-                favoriteFactory.deleteFromFavorites(index);
-                $scope.shouldShowDelete = false;
 
-        }
+                    var confirmPopup = $ionicPopup.confirm({
+                        title: 'Confirm Delete',
+                        template: 'Are you sure you want to delete this item?'
+                    });
+
+                    confirmPopup.then(function (res) {
+                        if (res) {
+                            console.log('Ok to delete');
+                            favoriteFactory.deleteFromFavorites(index);
+                        } else {
+                            console.log('Canceled delete');
+                        }
+                    });
+
+                    $scope.shouldShowDelete = false;
+
+            }
+
     }])
